+++ src/spicelib/devices/csw/cswnoise.c
@@ -93,7 +93,7 @@ if (!data->namelist) return(E_NOMEM);
 		switch (mode) {
 
 		case N_DENS:
-		    current_state = *(ckt->CKTstate0 + inst->CSWstate);
+		    current_state = (int)*(ckt->CKTstate0 + inst->CSWstate);
 		    NevalSrc(&noizDens,&lnNdens,ckt,THERMNOISE,
 				 inst->CSWposNode,inst->CSWnegNode,
 				 current_state?(model->CSWonConduct):(model->CSWoffConduct));
